+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Change Log
 
+## [0.9.2] - 2016-09-07
+- added flush to server prints so that buffered stdout on cygwin would work
+- added -T option to enable returning fake etags on certain APIs -instead of always doing it
+- added -D <dir>  option  where <dir> is the abs or relative path to the mockup.  if no -D option, then CWD is assumed
+
 ## [0.9.1] - 2016-09-06
 - Initial Public Release
 -